+++ swim_agent/src/lanes/mod.rs
@@ -12,5 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod command;
 pub mod map;
 pub mod value;
